Add files via upload
The data was collected from Data World and it was uploaded by @sparklesquad
Data for this project was generated using Synthea, a synthetic patient generator that models the medical history of patients. with the mission to output realistic quality patient data and associated health records covering every aspect of healthcare. The resulting data is free from cost, privacy, and security restrictions, enabling research with Health IT data that is otherwise legally or practically unavailable. De-identified real data still presents a challenge in the medical field because there are people who excel at re-identification of these data. For that reason the average medical center, etc. will not share their patient data. Most governmental data is at the hospital level. NHANES data is an exception.

I have tried to use basic to advanced SQL functions to get some insights from data. This is very helpful dataset if one is looking to practice healthcare data.
Thank you Synthea and https://data.world/sparklesquad for making this dataset available.